test program working, will overhaul functions for future simplicity

+++ src/components/programs/calculator/calculator.c
@@ -10,27 +10,27 @@ double add(double a, double b) {
 }
 
 EMSCRIPTEN_KEEPALIVE
-int subtract(int a, int b) {
+double subtract(double a, double b) {
   return a - b;
 }
 
 EMSCRIPTEN_KEEPALIVE
-int multiply(int a, int b) {
+double multiply(double a, double b) {
   return a * b;
 }
 
 EMSCRIPTEN_KEEPALIVE
-int divide(int a, int b) {
+double divide(double a, double b) {
   return a / b;
 }
 
 EMSCRIPTEN_KEEPALIVE
-int find_root(int root, int num) {
+double find_root(double root, double num) {
   return num + root; // TODO obviously not done, just init setup
 }
 
 EMSCRIPTEN_KEEPALIVE
-int percent(int a) {
+double percent(double a) {
   return a; // TODO find out exactly what this function should do based on how it works on win95
 }
